Client - Inplementando mais algumas funçoes e carrinho

+++ frontend/public/js/controllers/livro-controller.js
@@ -18,41 +18,54 @@ angular.module('livbri').controller('LivroController',function($scope,$routePara
     }
 
     $scope.carrinho =(livro)=>{
-        console.log(livro);
-        const titulo = livro.titulo;
-        const preco = livro.preco;
-        //Alert
-        Swal.fire({
-          title: 'Livro Adicionado Com Sucesso!',
-          text: "Quer Ver o Seu Carrinho?",
-          icon: 'success',
-          showCancelButton: true,
-          confirmButtonColor: '#3085d6',
-          cancelButtonColor: '#d33',
-          confirmButtonText: 'Sim!'
-        }).then((result) => {
-          if (result.value) {
-            let carrinho = localStorage.getItem('Livros');
-            carrinho.forEach(()=>{
-              console.log('este')
-            })
-            Swal.fire(
-              'Livros No Carrinho',
-              valores,
-              'success'
-            )
-          }
-        })
+      
+
+        
           
-          let livros = [];
+        let livros = [];
        
         livros = JSON.parse(localStorage.getItem('Livros')) || [];
-        // Inserindo um novo valor no localstorage
-        livros.push(livro.titulo);
-        // transformando em um object
-        localStorage.setItem('Livros', JSON.stringify(livros));
-            
-           
+        let mapLivros = livros.map(livro=>{
+          return livro.titulo;
+        }) 
+        const localizaLivros = mapLivros.indexOf(livro.titulo);
+        //verificando livros duplicados
+        if(localizaLivros != -1){
+            //se localizar nao insere o valor 
+            Swal.fire(
+              'Livro Localizado',
+              'Você ja adicionou esse livro no seu carrinho',
+              'warning'
+            )
+        }else{
+            //se nao localizado , insere o valor
+            // Inserindo um novo valor no localstorage
+            livros.push({titulo:livro.titulo,valor:livro.preco});
+            // transformando em um object
+            localStorage.setItem('Livros', JSON.stringify(livros));
+            //Alert
+            Swal.fire({
+              title: 'Livro Adicionado Com Sucesso!',
+              text: "Quer Ver o Seu Carrinho?",
+              icon: 'success',
+              showCancelButton: true,
+              confirmButtonColor: '#3085d6',
+              cancelButtonColor: '#d33',
+              confirmButtonText: 'Sim!'
+            }).then((result) => {
+              if (result.value) {
+                let carrinho = localStorage.getItem('Livros');
+                Swal.fire(
+                  'Livros No Carrinho',
+                  mapLivros+' '+livro.titulo+'<br>',
+                  'success'
+                )
+              }
+            })
+        }
+        
+
+        
         }
